requirements.txt
aiogram==3.22.0
aiosqlite==0.21.0
SQLAlchemy==2.0.44
python-dotenv==1.1.1
greenlet==3.0.3